fix: restore organization notification automation suite
